requirements.py3.10.txt
nltk
numpy
torch
spacy
attrs
allennlp
